+++ test/test_league_loader/test_FleaflickerLeagueLoader.py
@@ -4,7 +4,11 @@ from leeger.league_loader.FleaflickerLeagueLoader import FleaflickerLeagueLoader
 
 
 class TestFleaflickerLeagueLoader(unittest.TestCase):
+    """
+    # TODO: Add better tests.
+    """
+
     def test_noYearsGiven(self):
         with self.assertRaises(ValueError) as context:
-            leagueLoader = FleaflickerLeagueLoader("0", [])
+            FleaflickerLeagueLoader("0", [])
         self.assertEqual("No years given to load league with ID '0'.", str(context.exception))