+++ test/test_league_loader/test_MyFantasyLeagueLeagueLoader.py
@@ -6,6 +6,10 @@ from leeger.league_loader import MyFantasyLeagueLeagueLoader
 
 
 class TestMyFantasyLeagueLeagueLoader(unittest.TestCase):
+    """
+    # TODO: Add better tests.
+    """
+
     def test_loadLeague_intendedFailure(self):
         with self.assertRaises(MFLAPIClientException) as context:
             leagueLoader = MyFantasyLeagueLeagueLoader(
@@ -16,7 +20,7 @@ class TestMyFantasyLeagueLeagueLoader(unittest.TestCase):
 
     def test_noYearsGiven(self):
         with self.assertRaises(ValueError) as context:
-            leagueLoader = MyFantasyLeagueLeagueLoader(
+            MyFantasyLeagueLeagueLoader(
                 "0", [], mflUsername="", mflPassword="", mflUserAgentName=""
             )
         self.assertEqual("No years given to load league with ID '0'.", str(context.exception))